+++ packages/adapter-static/test/utils.js
@@ -4,7 +4,7 @@ import http from 'http';
 import { fileURLToPath } from 'url';
 import * as ports from 'port-authority';
 import sirv from 'sirv';
-import { chromium } from 'playwright-chromium';
+import { chromium } from '@playwright/test';
 import * as uvu from 'uvu';
 
 /**
@@ -48,7 +48,11 @@ export function run(app, callback) {
 			context.server = await create_server(context.port, handler);
 
 			context.base = `http://localhost:${context.port}`;
-			context.browser = await chromium.launch();
+			context.browser = await chromium.launch({
+				// use stable chrome from host OS instead of downloading one
+				// see https://playwright.dev/docs/browsers#google-chrome--microsoft-edge
+				channel: 'chrome'
+			});
 			context.page = await context.browser.newPage();
 		} catch (e) {
 			// TODO remove unnecessary try-catch https://github.com/lukeed/uvu/pull/61