+++ packages/kit/src/runtime/server/index.js
@@ -49,7 +49,7 @@ export async function respond(request, options, state) {
 	/** @type {Record<string, string>} */
 	let params = {};
 
-	if (options.paths.base && !state.prerender?.fallback) {
+	if (options.paths.base && !state.prerendering?.fallback) {
 		if (!decoded.startsWith(options.paths.base)) {
 			return new Response(undefined, { status: 404 });
 		}
@@ -60,16 +60,10 @@ export async function respond(request, options, state) {
 
 	if (is_data_request) {
 		decoded = decoded.slice(0, -DATA_SUFFIX.length) || '/';
-
-		const normalized = normalize_path(
-			url.pathname.slice(0, -DATA_SUFFIX.length),
-			options.trailing_slash
-		);
-
-		url = new URL(url.origin + normalized + url.search);
+		url = new URL(url.origin + url.pathname.slice(0, -DATA_SUFFIX.length) + url.search);
 	}
 
-	if (!state.prerender || !state.prerender.fallback) {
+	if (!state.prerendering?.fallback) {
 		const matchers = await options.manifest._.matchers();
 
 		for (const candidate of options.manifest._.routes) {
@@ -88,10 +82,11 @@ export async function respond(request, options, state) {
 	if (route?.type === 'page') {
 		const normalized = normalize_path(url.pathname, options.trailing_slash);
 
-		if (normalized !== url.pathname && !state.prerender?.fallback) {
+		if (normalized !== url.pathname && !state.prerendering?.fallback) {
 			return new Response(undefined, {
 				status: 301,
 				headers: {
+					'x-sveltekit-normalize': '1',
 					location:
 						// ensure paths starting with '//' are not treated as protocol-relative
 						(normalized.startsWith('//') ? url.origin + normalized : normalized) +
@@ -178,7 +173,7 @@ export async function respond(request, options, state) {
 					};
 				}
 
-				if (state.prerender && state.prerender.fallback) {
+				if (state.prerendering?.fallback) {
 					return await render_response({
 						event,
 						options,
@@ -280,7 +275,7 @@ export async function respond(request, options, state) {
 					});
 				}
 
-				if (state.prerender) {
+				if (state.prerendering) {
 					return new Response('not found', { status: 404 });
 				}