+++ packages/kit/test/prerendering/trailing-slash/svelte.config.js
@@ -1,0 +1,31 @@
+import path from 'path';
+import adapter from '../../../../adapter-static/index.js';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	kit: {
+		adapter: adapter({
+			fallback: '200.html'
+		}),
+
+		prerender: {
+			default: true
+		},
+
+		trailingSlash: 'always',
+
+		vite: {
+			build: {
+				minify: false
+			},
+			clearScreen: false,
+			server: {
+				fs: {
+					allow: [path.resolve('../../../src')]
+				}
+			}
+		}
+	}
+};
+
+export default config;