+++ packages/adapter-cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sveltejs/adapter-cloudflare-workers
 
+## 1.0.0-next.40
+
+### Patch Changes
+
+- [breaking] support Wrangler 2, drop Wrangler 1 ([#4887](https://github.com/sveltejs/kit/pull/4887))
+
 ## 1.0.0-next.39
 
 ### Patch Changes