+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,49 @@
 # create-svelte
 
+## 2.0.0-next.139
+
+### Patch Changes
+
+- correct default package.json format ([#5013](https://github.com/sveltejs/kit/pull/5013))
+
+* [breaking] Replace `%svelte.body%` with `%sveltekit.body%`, etc. ([#5016](https://github.com/sveltejs/kit/pull/5016))
+
+## 2.0.0-next.138
+
+### Patch Changes
+
+- Use separate ignore files for prettier and eslint ([#5009](https://github.com/sveltejs/kit/pull/5009))
+
+## 2.0.0-next.137
+
+### Patch Changes
+
+- import generated types from `__types/index.d.ts` file ([#4705](https://github.com/sveltejs/kit/pull/4705))
+
+## 2.0.0-next.136
+
+### Patch Changes
+
+- Add README ([#4951](https://github.com/sveltejs/kit/pull/4951))
+
+## 2.0.0-next.135
+
+### Patch Changes
+
+- remove unnecessary CHANGELOG.md ([#4903](https://github.com/sveltejs/kit/pull/4903))
+
+## 2.0.0-next.134
+
+### Patch Changes
+
+- fix lib, module, and target to not override the tsconfig they generate by default ([#4893](https://github.com/sveltejs/kit/pull/4893))
+
+## 2.0.0-next.133
+
+### Patch Changes
+
+- Bump `eslint` from version 7 to 8 ([#4553](https://github.com/sveltejs/kit/pull/4553))
+
 ## 2.0.0-next.132
 
 ### Patch Changes